split commoncrypto backend - hashes
